+++ src/presentation/view/VueStatistique.java
@@ -1,5 +1,6 @@
 package presentation.view;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.Map;
 
@@ -19,18 +20,18 @@ public class VueStatistique extends VueConsole {
         return lireDate("Date de fin (JJ/MM/AAAA) : ");
     }
 
-    public void afficherTotauxParCategorie(Map<Categorie, Double> totaux) {
+    public void afficherTotauxParCategorie(Map<Categorie, BigDecimal> totaux) {
         afficherMessage("Total dépensé par catégorie :");
         if (totaux.isEmpty()) {
             afficherMessage("  Aucune dépense sur cette période.");
             return;
         }
-        for (Map.Entry<Categorie, Double> entree : totaux.entrySet()) {
+        for (Map.Entry<Categorie, BigDecimal> entree : totaux.entrySet()) {
             afficherMessage(String.format("  %s : %.2f FCFA", entree.getKey().getLibelle(), entree.getValue()));
         }
     }
 
-    public void afficherTotalRevenusEtDepenses(double totalRevenus, double totalDepenses) {
+    public void afficherTotalRevenusEtDepenses(BigDecimal totalRevenus, BigDecimal totalDepenses) {
         afficherMessage(String.format("Total des revenus : %.2f FCFA", totalRevenus));
         afficherMessage(String.format("Total des dépenses : %.2f FCFA", totalDepenses));
     }
